requirements.txt
tushare>=1.4.2
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
loguru>=0.7.0
tabulate>=0.9.0
